requirements.txt
discord.py==2.0.1
discord_webhook==0.17.0
tweepy==4.10.1
